+++ packages/brick-scripts/src/questions/askTargetType.ts
@@ -16,15 +16,16 @@ export function askTargetType(): inquirer.ListQuestion<{
       TargetType.A_NEW_PACKAGE_OF_LIBS,
       TargetType.A_NEW_PACKAGE_OF_MICRO_APPS,
       TargetType.A_NEW_CUSTOM_PROVIDER_BRICK,
+      TargetType.A_NEW_CUSTOM_PROCESSOR,
       TargetType.A_NEW_PACKAGE_OF_PROVIDERS,
       TargetType.A_NEW_LEGACY_TEMPLATE,
       TargetType.A_NEW_PACKAGE_OF_LEGACY_TEMPLATES,
       TargetType.A_NEW_PACKAGE_OF_DLL,
       TargetType.TRANSFORM_A_MICRO_APP,
-      TargetType.I18N_PATCH_A_PACKAGE_OF_LEGACY_TEMPLATES
-    ].map(value => ({
+      TargetType.I18N_PATCH_A_PACKAGE_OF_LEGACY_TEMPLATES,
+    ].map((value) => ({
       name: TargetTypeDisplay[value],
-      value
-    }))
+      value,
+    })),
   };
 }