+++ packages/brick-scripts/src/questions/askTemplateName.ts
@@ -6,7 +6,7 @@ import { TargetType } from "../interface";
 export function askTemplateName({
   targetType,
   packageName,
-  appRoot
+  appRoot,
 }: {
   targetType: TargetType;
   packageName: string;
@@ -16,7 +16,9 @@ export function askTemplateName({
     type: "input",
     name: "templateName",
     message: `What's the name of your ${
-      targetType === TargetType.A_NEW_PACKAGE_OF_BRICKS ? "first" : "new"
+      targetType === TargetType.A_NEW_PACKAGE_OF_LEGACY_TEMPLATES
+        ? "first"
+        : "new"
     } template (in lower-kebab-case)?`,
     validate(value) {
       const pass = /^[a-z][a-z0-9]*(-[a-z0-9]+)*$/.test(value);
@@ -36,6 +38,6 @@ export function askTemplateName({
       }
 
       return true;
-    }
+    },
   };
 }