+++ packages/brick-scripts/template/custom-processor/$camelProcessorName$.spec.ts
@@ -1,0 +1,13 @@
+import { $camelProcessorName$ } from "./$camelProcessorName$";
+
+jest.mock("@easyops/brick-kit", () => ({
+  getRuntime: () => ({
+    registerCustomProcessor: jest.fn(),
+  }),
+}));
+
+describe("$camelProcessorName$", () => {
+  it("should work", () => {
+    expect($camelProcessorName$()).toEqual(undefined);
+  });
+});