+++ packages/brick-utils/src/cook/index.ts
@@ -1,2 +1,3 @@
 export { cook } from "./cook";
 export { precook } from "./precook";
+export * from "./preevaluate";