+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.5...@easyops/build-config-factory@1.18.6) (2020-08-10)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [1.18.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.4...@easyops/build-config-factory@1.18.5) (2020-08-07)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [1.18.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.3...@easyops/build-config-factory@1.18.4) (2020-08-07)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [1.18.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.2...@easyops/build-config-factory@1.18.3) (2020-08-05)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
 # [2.0.0-next.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.2...@easyops/build-config-factory@2.0.0-next.0) (2020-08-05)
 
 **Note:** Version bump only for package @easyops/build-config-factory