+++ packages/create-next-repo/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.34](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.33...@easyops/create-next-repo@0.13.34) (2020-08-10)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.33](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.32...@easyops/create-next-repo@0.13.33) (2020-08-07)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.32](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.31...@easyops/create-next-repo@0.13.32) (2020-08-07)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.31](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.30...@easyops/create-next-repo@0.13.31) (2020-08-07)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.30](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.29...@easyops/create-next-repo@0.13.30) (2020-08-05)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.29](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.28...@easyops/create-next-repo@0.13.29) (2020-08-05)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
 # [1.0.0-next.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.0-next.1...@easyops/create-next-repo@1.0.0-next.2) (2020-08-05)
 
 **Note:** Version bump only for package @easyops/create-next-repo